# Conflicts:
#	kwdbts2/ts_engine/ts_compressor_impl.cpp

+++ kwdbts2/ts_engine/ts_bloomfilter.cpp
@@ -84,7 +84,7 @@ void TsBloomFilter::CalcHash(TSEntityID eid, std::vector<size_t>* result) const 
   assert(result->size() == k_);
   auto h1 = hash1(eid);
   auto h2 = hash2(eid);
-  for (int i = 0; i < k_; ++i) {
+  for (size_t i = 0; i < k_; ++i) {
     size_t hash = h1 + i * h2 + i * i * i;
     size_t n = hash % m_;
     (*result)[i] = n;
